[optimize] 建立type資料夾並集中所有共用的ts，優化部分程式碼寫法
