+++ src/api/user.tsx
@@ -1,32 +1,13 @@
 import axios from 'axios';
 import { API_URL } from './index';
-import { ResponseType } from '../types/api';
+import { AxResponseType } from '../types/apiType';
+import { UserDataType } from '../types/userType';
 
 const baseUrl = API_URL;
 const authToken = localStorage.getItem('authToken');
 
-/** 使用者資料型別 */
-export interface UserDataType {
-  userId: string;
-  email: string;
-  account: string;
-  name: string;
-  avatar: string;
-  bio: string;
-  userRole: number;
-  createdAt: string;
-  status: number;
-}
 
-export interface UserSettingType {
-  language: string;
-  theme: number;
-  tags: [string];
-  emailPrompt: Boolean;
-  mobilePrompt: Boolean;
-}
-
-interface GetUserProfileType extends ResponseType {
+interface GetUserProfileType extends AxResponseType {
   data: UserDataType;
 }
 
@@ -43,7 +24,6 @@ export async function getUserProfile(userId: string): Promise<GetUserProfileType
     })
     .catch((error) => {
       return error;
-      // throw new Error(error.response?.data?.message || '取得使用者資料時發生錯誤');
     });
   return result;
 }
