+++ src/components/user/UserInfoPanel.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
 // --- components ---
 import Avatar from './Avatar';
-import FollowBtn from './FollowBtn';
 
-function AuthorInfoPanel(props: { account: string; name: string; avatarUrl: string }) {
+function UserInfoPanel(props: { account: string; name: string; avatarUrl: string }) {
   const { account, name, avatarUrl } = props;
   return (
     <div className="flex items-center my-4">
       <div className="flex justify-center items-center mr-4">
-        <Avatar avatarUrl={avatarUrl} size="w-11 h-11" textSize="text-xl" />
+        <Avatar name={name} avatarUrl={avatarUrl} size="w-11 h-11" textSize="text-xl" />
       </div>
       <div>
         <p className="font-semibold">{name}</p>
@@ -18,4 +17,4 @@ function AuthorInfoPanel(props: { account: string; name: string; avatarUrl: stri
   );
 }
 
-export default AuthorInfoPanel;
+export default UserInfoPanel;
