--- src/types/api.ts
@@ -1,9 +1,0 @@
-/** API Response 通用型別 */
-export interface ResponseType {
-  config: object;
-  data: object;
-  headers: object;
-  request: object;
-  response: object;
-  status: number;
-}
