+++ src/types/articleType.ts
@@ -1,0 +1,10 @@
+
+/** articleList 型別 */
+export interface ArticleListType {
+  body: string;
+  id: number;
+  reactions: number;
+  tags: string[];
+  title: string;
+  userId: number;
+}
