+++ src/types/authType.ts
@@ -1,0 +1,12 @@
+/** 註冊參數型別 */
+export interface SignUpParamType {
+  email: string;
+  password: string;
+  confirmPassword: string;
+}
+
+/** 登入參數型別 */
+export interface SignInParamType {
+  email: string;
+  password: string;
+}
